lifecycle event with class component and functional comonent exmaple added
